% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consulta_enaho.R
\name{consulta_enaho}
\alias{consulta_enaho}
\title{consulta_enaho}
\usage{
consulta_enaho(periodo, codigo_modulo, base, guardar = FALSE,
  ruta = "", codificacion = NULL)
}
\arguments{
\item{periodo}{A침o de la encuesta}

\item{codigo_modulo}{C칩digo del m칩dulo}

\item{base}{Nombre de base de datos}

\item{guardar}{Logical}

\item{ruta}{Ruta para guardar archivo}

\item{codificacion}{Codificaci칩n (UTF-8, latin1, etc.)}
}
\value{
archivo .sav o Data Frame (Segun parametro guardar)
}
\description{
Funcion para extraer bases de datos de la Encuesta Nacional de Hogares.
}
\examples{
consulta_enaho(periodo = 2017, codigo_modulo = "01", base = "Enaho01-2017-100", guardar = FALSE)

}
