% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consulta_endes.R
\name{consulta_endes}
\alias{consulta_endes}
\title{consulta_endes}
\usage{
consulta_endes(periodo, codigo_modulo, base, guardar = FALSE,
  ruta = "", codificacion = NULL)
}
\arguments{
\item{periodo}{A침o de la encuesta}

\item{codigo_modulo}{C칩digo del m칩dulo}

\item{base}{Nombre de base de datos}

\item{guardar}{Logical}

\item{ruta}{Ruta para guardar archivo}

\item{codificacion}{Codificaci칩n (UTF-8, latin1, etc.)}
}
\value{
archivo .sav o Data Frame (Segun parametro guardar)
}
\description{
Funcion para extraer bases de datos de la Encuesta Nacional Demografica y de Salud familiar.
}
\examples{
consulta_endes(periodo = 2012,codigo_modulo = '64',base = 'RECH1', 
guardar = FALSE, codificacion = "UTF-8")

}
