% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{data}
\alias{data}
\title{Prices of 50,000 round cut diamonds}
\format{
Un data frame con 212 observaciones y 4 variables
}
\usage{
data(modulos_endes)
}
\description{
Este dataset contiene la lista de modulos y a침os de encuestas ENDES y ENAHO como referencia.
}
\details{
\itemize{
  \item periodo. periodo de la encuesta (2020-2000)
  \item nombre_encuesta. nombre de la encuesta(ENDES, ENAHO)
  \item codigo_modulo. codigo del m칩dulo de la encuesta
  \item modulo. descripci칩n del m칩dulo (worst) to D (best)
}
}
\keyword{datasets}
