% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionario_encuesta.R
\name{diccionario_encuesta}
\alias{diccionario_encuesta}
\title{diccionario_encuesta}
\usage{
diccionario_encuesta(base_endes)
}
\arguments{
\item{base_endes}{Base Endes o Enaho}
}
\value{
tibble
}
\description{
Funcion para generar diccionario de variables con las bases de datos de la ENDES o ENAHO.
}
\examples{

endes <- consulta_endes(periodo = 2012,codigo_modulo = '64',
base = 'RECH1', guardar = FALSE, codificacion = "UTF-8")
diccionario_encuesta(base_endes=endes)

enaho <- consulta_enaho(periodo = 2017, codigo_modulo = "01", 
base = "Enaho01-2017-100", guardar = FALSE, codificacion = "latin1")

diccionario_encuesta(base_endes=enaho)

}
