% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unir_endes.R
\name{unir_endes}
\alias{unir_endes}
\title{unir_endes}
\usage{
unir_endes(base1, base2, tipo_union = "individual")
}
\arguments{
\item{base1, base2, tipo_union}{}
}
\value{
Data.Frame
}
\description{
Funcion para unir bases de datos de la ENDES
}
\examples{

}
