% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivos_enaho.R
\name{archivos_enaho}
\alias{archivos_enaho}
\title{archivos_enaho}
\usage{
archivos_enaho(periodo, codigo_modulo)
}
\arguments{
\item{periodo}{A침o de la encuesta}

\item{codigo_modulo}{C칩digo del m칩dulo}
}
\value{
tibble
}
\description{
Funcion para listar archivos disponibles en cada m칩dulo por periodo de la base de datos de la ENDES o ENAHO.
}
\examples{

archivos_enaho(periodo=2017, codigo_modulo="01")

}
