% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consulta_enaho.R
\name{consulta_enaho}
\alias{consulta_enaho}
\title{consulta_enaho}
\usage{
consulta_enaho(periodo, codigo_modulo, base, guardar = FALSE,
  ruta = "", codificacion = NULL)
}
\arguments{
\item{periodo, codigo_modulo, base, guardar, ruta, codificacion}{}
}
\value{
archivo .sav o Data Frame (Segun parametro guardar)
}
\description{
Funcion para extraer bases de datos de la Encuesta Nacional de Hogares.
}
\examples{
consulta_enaho(periodo = 2017, codigo_modulo = "01", base = "Enaho01-2017-100", guardar = FALSE)

}
