% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consulta_endes.R
\name{consulta_endes}
\alias{consulta_endes}
\title{consulta_endes}
\usage{
consulta_endes(periodo, codigo_modulo, base, guardar = TRUE, ruta)
}
\arguments{
\item{periodo, codigo_modulo, base, guardar, ruta}{}
}
\value{
archivo .sav o Data Frame (Segun parametro guardar)
}
\description{
Funcion para extraer bases de datos de la Encuesta Nacional Demografica y de Salud familiar.
}
\examples{
consulta_endes(periodo = 2012,codigo_modulo = '64',base = 'RECH1', guardar = F)

}
