% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unir_endes.R
\name{unir_endes}
\alias{unir_endes}
\title{unir_endes}
\usage{
unir_endes(base1, base2, tipo_union = "individual")
}
\arguments{
\item{base1}{Base Endes o Enaho}

\item{base2}{Base Endes o Enaho}

\item{tipo_union}{Tipo de unión (Hogar o Individual)}
}
\value{
Data.Frame
}
\description{
Funcion para unir bases de datos de la ENDES
}
\examples{

persona   <- consulta_endes(periodo = 2017, codigo_modulo = 64, 
base = "RECH1", guardar = FALSE, codificacion = "latin1")

salud     <- consulta_endes(periodo = 2017, codigo_modulo = 414, 
base = "CSALUD01", guardar = FALSE, codificacion = "latin1")

unir_endes(base1 = persona, base2 = salud, tipo_union = "individual")

}
